Split the NavTile factory function into it's own class in the 'NavMeshPlus.Editors.Extensions' namespace
