+++ NavMeshComponents/Editor/TilemapExtras/NavTileFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62e56a0d944a45b995bdd5ecd6d9caff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
